Deleted files can now be restored, renamed files can be undone, colour coded renames

+++ utils/file_operations.py
@@ -4,6 +4,9 @@ from utils.structures import Item
 
 
 def rename_file(before: Item, after: str):
+    location = before.path
+    if os.path.exists(location / after):
+        return False
     try:
         os.rename(before.location, before.path / after)
     except (FileNotFoundError, PermissionError):
@@ -17,6 +20,16 @@ def delete_file(item: Item):
             shutil.rmtree(item.location)
         else:
             os.remove(item.location)
+    except (FileNotFoundError, PermissionError):
+        return False
+    return True
+
+def create_file_with_content(location: str, content: str):
+    if os.path.exists(location):
+        return False
+    try:
+        with open(location, "w") as f:
+            f.write(content)
     except (FileNotFoundError, PermissionError):
         return False
     return True
